fix: add wiremock + move utils to core
