+++ core/Cargo.toml
@@ -25,3 +25,4 @@ mockall = "0.11.3"
 tokio = { version = "1.21.2", features = ["macros"] }
 serde_json = "1.0"
 serde = { version = "1.0" }
+wiremock = "0.5.17"
