+++ core/src/client/constants.rs
@@ -12,6 +12,6 @@ pub mod selectors {
     pub const BYTECODE: FieldElement = selector!("bytecode");
 
     pub const EXECUTE_AT_ADDRESS: FieldElement = selector!("execute_at_address");
-
     pub const COMPUTE_STARKNET_ADDRESS: FieldElement = selector!("compute_starknet_address");
+    pub const KKRT_CHAIN_ID: u64 = 1263227476_u64;
 }
