+++ core/src/utils/data/starknet_getBlockWithTxHashes.json
