+++ core/src/utils/data/starknet_getBlockWithTxs.json
