+++ core/src/utils/data/starknet_getCode.json
