+++ core/src/utils/wiremock_utils.rs
@@ -1,10 +1,6 @@
+use crate::client::{constants::selectors::GET_STARKNET_CONTRACT_ADDRESS, StarknetClientImpl};
 use jsonrpsee::server::ServerHandle;
-use kakarot_rpc::run_server;
-use kakarot_rpc_core::lightclient::{
-    constants::{selectors::GET_STARKNET_CONTRACT_ADDRESS, ACCOUNT_REGISTRY_ADDRESS},
-    StarknetClientImpl,
-};
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use starknet::{
     core::types::FieldElement,
     providers::jsonrpc::models::{BlockId, BlockTag, FunctionCall},
@@ -20,6 +16,13 @@ pub struct StarknetRpcBaseData<'a, StarknetParams> {
     jsonrpc: &'a str,
     method: &'a str,
     params: StarknetParams,
+}
+
+#[derive(Deserialize, Debug)]
+pub struct EthJsonRpcResponse<StarknetParams> {
+    pub id: usize,
+    pub jsonrpc: String,
+    pub result: StarknetParams,
 }
 
 #[derive(Serialize)]
@@ -70,7 +73,7 @@ impl<'a, StarknetParams> StarknetRpcBaseData<'a, StarknetParams> {
     }
 }
 
-async fn setup_wiremock() -> String {
+pub async fn setup_wiremock() -> String {
     let mock_server = MockServer::start().await;
     let empty_vec: Vec<&str> = Vec::new();
     Mock::given(method("POST"))
@@ -120,67 +123,34 @@ async fn setup_wiremock() -> String {
         .await;
 
     // Get kakarot contract addess
-    let tx_calldata_vec =
-        vec![FieldElement::from_hex_be("0xabde1007dcf45cb509ddde375162399a99880064").unwrap()];
-    let request = FunctionCall {
-        contract_address: FieldElement::from_hex_be(ACCOUNT_REGISTRY_ADDRESS).unwrap(),
-        entry_point_selector: GET_STARKNET_CONTRACT_ADDRESS,
-        calldata: tx_calldata_vec,
-    };
-    let block_id = BlockId::Tag(BlockTag::Latest);
-    Mock::given(method("POST"))
-        .and(body_json(StarknetRpcBaseData::call([
-            serde_json::to_value(request).unwrap(),
-            serde_json::to_value(block_id).unwrap(),
-        ])))
-        .respond_with(
-            ResponseTemplate::new(200)
-                .set_body_raw(
-                    include_str!("data/starknet_getCode.json"),
-                    "application/json",
-                )
-                .append_header("vary", "Accept-Encoding")
-                .append_header("vary", "Origin"),
-        )
-        .mount(&mock_server)
-        .await;
+    // TODO: Change this method with associated correct code
+    // let tx_calldata_vec =
+    //     vec![FieldElement::from_hex_be("0xabde1007dcf45cb509ddde375162399a99880064").unwrap()];
+    // let request = FunctionCall {
+    //     contract_address: FieldElement::from_hex_be(ACCOUNT_REGISTRY_ADDRESS).unwrap(),
+    //     entry_point_selector: GET_STARKNET_CONTRACT_ADDRESS,
+    //     calldata: tx_calldata_vec,
+    // };
+    // let block_id = BlockId::Tag(BlockTag::Latest);
+    // Mock::given(method("POST"))
+    //     .and(body_json(StarknetRpcBaseData::call([
+    //         serde_json::to_value(request).unwrap(),
+    //         serde_json::to_value(block_id).unwrap(),
+    //     ])))
+    //     .respond_with(
+    //         ResponseTemplate::new(200)
+    //             .set_body_raw(
+    //                 include_str!("data/starknet_getCode.json"),
+    //                 "application/json",
+    //             )
+    //             .append_header("vary", "Accept-Encoding")
+    //             .append_header("vary", "Origin"),
+    //     )
+    //     .mount(&mock_server)
+    //     .await;
 
     // Get kakarot contract bytecode
-    // todo!();
+    // TODO: Use the latest mapping between starknet and EVM adresses
 
     mock_server.uri()
-}
-
-/// Run wiremock to fake starknet rpc and then run our own kakarot_rpc_server.
-///
-/// Example :
-/// ```
-///   use kakarot_rpc_utils::wiremock_utils::setup_rpc_server;
-///
-///   #[tokio::test]
-///   async fn test_case() {
-///       // Run base server
-///       let (_, server_handle) = setup_rpc_server().await;
-///
-///       //Query whatever eth_rpc endpoints
-///       let client = reqwest::Client::new();
-///        let res = client
-///            .post("http://127.0.0.1:3030")
-///            .body("{\"jsonrpc\": \"2.0\", \"id\": 1, \"method\": \"eth_chainId\", \"params\": [] }")
-///            .header("content-type", "application/json")
-///            .send()
-///            .await
-///            .unwrap();
-///
-///        // Dont forget to close server at the end.
-///        let _has_stop = server_handle.stop().unwrap();
-///   }
-/// ```
-pub async fn setup_rpc_server() -> (String, ServerHandle) {
-    let starknet_rpc = setup_wiremock().await;
-
-    let starknet_lightclient = StarknetClientImpl::new(&starknet_rpc).unwrap();
-    let (_rpc_server_uri, server_handle) =
-        run_server(Box::new(starknet_lightclient)).await.unwrap();
-    (starknet_rpc, server_handle)
 }
