--- internal/utils/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "kakarot_rpc_utils"
-version = "0.1.0"
-edition.workspace = true
-description.workspace = true
-homepage.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-kakarot_rpc_core = { path = "../../core" }
-kakarot_rpc = { path = "../../rpc" }
-log = "0.4.17"
-
-# rpc
-tokio = { version = "1.21.2", features = ["macros"] }
-serde = { version = "1.0.147", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0.38"
-hex = "0.4"
-starknet = { git = "https://github.com/xJonathanLEI/starknet-rs" }
-reqwest = "0.11.13"
-wiremock = "0.5.17"
-jsonrpsee = { version = "^0.16", features=["full"] }
-reth-primitives = { git = "https://github.com/paradigmxyz/reth.git"}
