+++ rpc/Cargo.toml
@@ -36,9 +36,6 @@ dotenv = "0.15.0"
 mockall = "0.11.3"
 anyhow = "1.0.68"
 
-[dev-dependencies]
-kakarot_rpc_utils = { path = "../internal/utils" }
-
 [[bin]]
 name = "kakarot-rpc"
 path = "src/main.rs"
