+++ rpc/tests/rpc.rs
@@ -1,9 +1,14 @@
 #[cfg(test)]
-mod test {
+mod tests {
     use kakarot_rpc::eth_rpc::KakarotEthRpc;
+    use kakarot_rpc::test_utils::setup_rpc_server;
     use kakarot_rpc_core::client::MockStarknetClient;
     use reth_primitives::{Address, Bytes};
 
+    use kakarot_rpc_core::{
+        client::constants::selectors::KKRT_CHAIN_ID, utils::wiremock_utils::EthJsonRpcResponse,
+    };
+    use reth_primitives::{BlockHash, ChainId};
     use starknet::providers::jsonrpc::models::BlockId as StarknetBlockId;
 
     #[tokio::test]
@@ -18,77 +23,10 @@ mod test {
             .await
             .unwrap();
 
-        let _has_stop = server_handle.stop().unwrap();
+        let block_hash = res.json::<EthJsonRpcResponse<String>>().await.unwrap();
+        // see rpc/src/eth_rpc.rs:286 to get ASCII String
+        assert_eq!(block_hash.result, String::from("0x4b4b5254"));
 
-        // check rpc response match our needs
-        // for this request check hexa is KKRT
-    }
-
-    #[tokio::test]
-    async fn test_get_block_number_is_ok() {
-        let (_, server_handle) = setup_rpc_server().await;
-        let client = reqwest::Client::new();
-        let res = client
-            .post("http://127.0.0.1:3030")
-            .body("{\"jsonrpc\": \"2.0\", \"id\": 1, \"method\": \"eth_blockNumber\", \"params\": [] }")
-            .header("content-type", "application/json")
-            .send()
-            .await
-            .unwrap();
-
-        let _has_stop = server_handle.stop().unwrap();
-
-        // response should return a fixed block id for the moment eg. 19640
-    }
-
-    #[tokio::test]
-    async fn test_get_code_is_ok() {
-        let (_, server_handle) = setup_rpc_server().await;
-        let client = reqwest::Client::new();
-        let res = client
-            .post("http://127.0.0.1:3030")
-            .body("{\"jsonrpc\": \"2.0\", \"id\": 1, \"method\": \"eth_getCode\", \"params\": [\"0xabde1007dcf45cb509ddde375162399a99880064\", \"latest\"] }")
-            .header("content-type", "application/json")
-            .send()
-            .await
-            .unwrap();
-
-        let _has_stop = server_handle.stop().unwrap();
-
-        // response should return a fixed block id for the moment eg. 19640
-    }
-
-    #[tokio::test]
-    async fn test_block_by_number_is_ok() {
-        let (_, server_handle) = setup_rpc_server().await;
-        let client = reqwest::Client::new();
-        let res = client
-            .post("http://127.0.0.1:3030")
-            .body("{\"jsonrpc\": \"2.0\", \"id\": 1, \"method\": \"eth_getBlockByNumber\", \"params\": [] }")
-            .header("content-type", "application/json")
-            .send()
-            .await
-            .unwrap();
-
-        let _has_stop = server_handle.stop().unwrap();
-
-        // response should return a fixed block id for the moment eg. 19640
-    }
-
-    #[tokio::test]
-    async fn test_block_by_hash_is_ok() {
-        let (_, server_handle) = setup_rpc_server().await;
-        let client = reqwest::Client::new();
-        let res = client
-            .post("http://127.0.0.1:3030")
-            .body("{\"jsonrpc\": \"2.0\", \"id\": 1, \"method\": \"eth_getBlockByHash\", \"params\": [] }")
-            .header("content-type", "application/json")
-            .send()
-            .await
-            .unwrap();
-
-        let _has_stop = server_handle.stop().unwrap();
-
-        // response should return a fixed block id for the moment eg. 19640
+        server_handle.stop().unwrap();
     }
 }
